separate extractor from database
